create component userInfo
