+++ my-app/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\index.tsx":"1","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\reportWebVitals.ts":"2","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\App.tsx":"3","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Sidebar\\Sidebar.tsx":"4","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Sidebar\\Brand.tsx":"5","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Sidebar\\UserInfo.tsx":"6"},{"size":151,"mtime":1613503395069,"results":"7","hashOfConfig":"8"},{"size":425,"mtime":499162500000,"results":"9","hashOfConfig":"8"},{"size":215,"mtime":1613581723831,"results":"10","hashOfConfig":"8"},{"size":303,"mtime":1613582812883,"results":"11","hashOfConfig":"8"},{"size":513,"mtime":1613582666589,"results":"12","hashOfConfig":"8"},{"size":543,"mtime":1613583259465,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1oa2vh",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"19"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\index.tsx",[],["28","29"],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\reportWebVitals.ts",[],["30","31"],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\App.tsx",[],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Sidebar\\Sidebar.tsx",[],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Sidebar\\Brand.tsx",[],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Sidebar\\UserInfo.tsx",[],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"32","replacedBy":"36"},{"ruleId":"34","replacedBy":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],["38"],["39"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\index.tsx":"1","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\reportWebVitals.ts":"2","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\App.tsx":"3","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Components\\Sidebar.tsx":"4","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Components\\Brand.tsx":"5","C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Components\\UserInfo.tsx":"6"},{"size":151,"mtime":1613586569016,"results":"7","hashOfConfig":"8"},{"size":425,"mtime":499162500000,"results":"9","hashOfConfig":"8"},{"size":240,"mtime":1613649444938,"results":"10","hashOfConfig":"8"},{"size":567,"mtime":1613649249862,"results":"11","hashOfConfig":"8"},{"size":513,"mtime":1613582666589,"results":"12","hashOfConfig":"8"},{"size":659,"mtime":1613649217403,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1oa2vh",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"19"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\index.tsx",[],["28","29"],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\reportWebVitals.ts",[],["30","31"],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\App.tsx",[],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Components\\Sidebar.tsx",[],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Components\\Brand.tsx",[],"C:\\Users\\Mariya\\Desktop\\dev\\FrontendLearning-\\my-app\\src\\Components\\UserInfo.tsx",[],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"32","replacedBy":"36"},{"ruleId":"34","replacedBy":"37"},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],["38"],["39"],"no-global-assign","no-unsafe-negation"]
